* fuzz/*.in/*: Update fuzzer corpora

+++ fuzz/libwget_bar_fuzzer.c
@@ -47,7 +47,7 @@ int LLVMFuzzerTestOneInput(const uint8_t *data, size_t size)
 	wget_bar_set_slots(bar, 2);
 
 	wget_bar_slot_begin(bar, 1, "test", 64000);
-	wget_bar_slot_downloaded(bar, 1, atoi(in));
+	wget_bar_slot_downloaded(bar, 1, (unsigned) atoi(in));
 	wget_bar_printf(bar, 1, "%s", in);
 	wget_bar_write_line(bar, (char *) data, size);
 	wget_bar_update(bar);
